Devise now uses username instead of email
User.rb was changed to how easy ways of accessing login, and checks
formatting of username.

Had to add all of the devise views so they could be changed to use login
instead of email.

init/devise.rb had to be changed to use login

application_controller had to be updated as well.

A migration was made to add username to the table for Users.