account disabling and cleaning
Pages were refreshing SUPER slow because of all the unnecessary css and
js files being loaded. We still have them, but they are in a backup
folder incase we want to use them later.

Accounts can now be destroyed if there are no transactions.

Accounts can now be disabled and enabled. The function does nothing
except switch the bool is_active.

Put the tables for transactions and accounts into partials instead of
having Austin use render :template. This is due to I want the index
functions to be administrator/teller functions with search options so we
can dynamically load the table.