Applied currency changes to rest of site.
Only had to update the account and transaction models, for the class
side of the update.

I had to go to each controller and update the strong params to use the
new string instead of the decimal.

Updated each form to use the new string instead of decimal.

updated the en.yml to give proper feedback for error messages.