Bill Pay
This is the bill pay update

There is a bill function called pay_all that will go through every bill
that needs to be paid that day and transfer the funds.

Modified validations for accounts. The numerality is is different for
the different types so I have individual checks for each class instead
of the global check all.